refactor: onboarding と settings を共通化
